feat: add Type::Float as independent type (step 1 of float support)
Add Float variant to Type enum:
- sizeof(float) = 4, alignment = 4
- Parser: `float` keyword now produces Type::Float (was Type::Double)
- PartialEq: Float == Float
- Codegen: temporary AsmType::Longword mapping (TODO: SSE single)
- is_float(), is_floating() predicates added

Remaining steps for full float support:
- Type checker: float↔double implicit promotion (usual arithmetic conversions)
- TACKY gen: IntToFloat, FloatToDouble, DoubleToFloat conversions
- Codegen: movss/addss/subss/mulss/divss (SSE single-precision)
- Emitter: ss suffix instructions
- Regalloc: XMM register handling for float
- Meaning-preservation tests

No existing tests affected (no corpus uses float-specific precision).

+++ src/parse/parser.rs
@@ -449,10 +449,10 @@ impl<'a> Parser<'a> {
                         self.advance()?;
                         return Ok(Type::Long);
                     }
-                    // float → Double として近似（parse-only）
+                    // float → 独立型（IEEE 754 single-precision）
                     if name == "float" {
                         self.advance()?;
-                        return Ok(Type::Double);
+                        return Ok(Type::Float);
                     }
                     // 他の型キーワードが未出現の場合のみ typedef 名として認識
                     if let Some(ty) = self.typedef_names.get(name).cloned() {
